Input   - added PM collapse tests   - fixed gravity_[cm]g axis="z" not "x" Refine   - added RefineParticle[Mass|Count]   - removed FieldDescr argument to Refine::apply()   - fixed FieldData::size() parameter error in RefineMask (need FieldDescr*)     or use Field   - updating RefineFoo to use Field not FieldData / FieldDescr Output   - finished support for OutputImage axis (was partially implemented)   - added OutputImage::image_abs_ (why?)   - removed OutputImage::nzi_   - removed OutputImage::set_axis() (axis passed into constructor instead) Memory   - renamed Memory::memory_allocate_warning_ -> max_allocate_warning   - renamed Memory::memory_max_allocated_error_ -> max_allocated_error Parameters   - added Output:<group>:image_abs parameter   - initialize Config::output_axis parameter Compute   - removed unnecessary EnzoComputeAcceleration:symmetric_ Initial   - added mask to InitialPm   - support uniform placement of particles on grid (centered in cells)     for EnzoInitialPm Method   - Added EnzoMethodPmDeposit::alpha_ for time t + alpha_*dt to support     t + 0.5*dt

+++ src/Cello/data_Data.hpp
@@ -100,11 +100,11 @@ public: // interface
   { return num_field_data_; }
 
   /// Return the ith Field data
-  const FieldData * field_data (int i=0) const throw()
+  const FieldData * field_data (size_t i=0) const throw()
   { return (i < field_data_.size()) ? field_data_[i] : NULL; }
 
   /// Return the ith Field data
-  FieldData * field_data (int i=0) throw()
+  FieldData * field_data (size_t i=0) throw()
   { return (i < field_data_.size()) ? field_data_[i] : NULL; }
 
   // Return the ith Field descriptor
